services/event-service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aerich[toml]==0.8.1
    # via -r requirements.in
aio-pika==9.6.1
    # via -r requirements.in
aiormq==6.9.3
    # via aio-pika
aiosqlite==0.22.1
    # via tortoise-orm
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   starlette
    #   tortoise-orm
asyncclick==8.3.0.7
    # via aerich
asyncpg==0.31.0
    # via
    #   -r requirements.in
    #   tortoise-orm
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   asyncclick
    #   click
dictdiffer==0.9.0
    # via aerich
fastapi==0.134.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   yarl
iso8601==2.1.0
    # via tortoise-orm
multidict==6.7.1
    # via yarl
pamqp==3.3.0
    # via aiormq
propcache==0.4.1
    # via yarl
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   aerich
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r requirements.in
pyjwt==2.11.0
    # via -r requirements.in
pypika-tortoise==0.6.3
    # via tortoise-orm
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
starlette==0.52.1
    # via fastapi
tomli-w==1.2.0
    # via aerich
tortoise-orm[asyncpg]==1.1.5
    # via
    #   -r requirements.in
    #   aerich
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
uvicorn==0.41.0
    # via -r requirements.in
yarl==1.22.0
    # via
    #   aio-pika
    #   aiormq
